requirements.txt
faiss-cpu==1.11.0
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
numpy==2.2.6
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3

pytest~=8.4.0
beautifulsoup4~=4.13.4
pydantic~=2.11.5